fix rating reservation
the query always returned 0 answers which led to the "never rate
anything bug"

the query before was very messed up, I wrote a new one which gets all
answers which are not rated more than the maximum time configured in the
experiment. Ordered by the count of the ratings per answer. so less
rated answeres are rated earlier. Or should this depend on quality?